requirements.txt
Django
djangorestframework
Flask
boto3
PyJWT
aws-cdk-lib
constructs
python-dotenv